requirements.txt
fastapi
uvicorn
requests
pdfplumber
faiss-cpu
